Move execution-related code into subpackage
