--- sources/src/execution.ts
@@ -1,17 +1,0 @@
-import * as core from '@actions/core'
-import * as exec from '@actions/exec'
-import * as gradlew from './gradlew'
-
-export async function executeGradleBuild(executable: string | undefined, root: string, args: string[]): Promise<void> {
-    // Use the provided executable, or look for a Gradle wrapper script to run
-    const toExecute = executable ?? gradlew.gradleWrapperScript(root)
-
-    const status: number = await exec.exec(toExecute, args, {
-        cwd: root,
-        ignoreReturnCode: true
-    })
-
-    if (status !== 0) {
-        core.setFailed(`Gradle build failed: see console output for details`)
-    }
-}
